Refactor `AgentEvent` into `SSEAgent.Event`, improving event hierarchy and integration with `SSEAgent`.
